requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-mistralai>=0.2.0
langchain-qdrant>=0.2.0
langchain-community>=0.3.0
qdrant-client>=1.16.0
neo4j>=5.14.0
streamlit>=1.29.0
pypdf>=3.17.0
pdf2image>=1.17.0
Pillow>=10.0.0
mistralai>=1.0.0
